move getChangedEntriesInReport() to a separate file, reuse it in both reporters
