+++ packages/bundle-size/src/reporters/cliReporter.test.js
@@ -26,11 +26,11 @@ describe('cliReporter', () => {
       "┌────────────────────┬────────┬───────────────────────┐
       │ Fixture            │ Before │ After (minified/GZIP) │
       ├────────────────────┼────────┼───────────────────────┤
-      │ foo-package        │    N/A │            100%↑ 1 kB │
-      │ New entry (new)    │    N/A │           100%↑ 100 B │
-      ├────────────────────┼────────┼───────────────────────┤
       │ baz-package        │   2 kB │            100%↑ 1 kB │
       │ An entry with diff │  200 B │           100%↑ 100 B │
+      ├────────────────────┼────────┼───────────────────────┤
+      │ foo-package        │    N/A │            100%↑ 1 kB │
+      │ New entry (new)    │    N/A │           100%↑ 100 B │
       └────────────────────┴────────┴───────────────────────┘"
     `);
   });
